+++ packages/talker_http_logger/lib/http_response_log.dart
@@ -22,7 +22,7 @@ class HttpResponseLog extends TalkerLog with ResponseTime {
   AnsiPen get pen => settings.responsePen ?? (AnsiPen()..xterm(46));
 
   @override
-  String get key => TalkerLogType.httpResponse.key;
+  String get key => TalkerKey.httpResponse;
 
   @override
   LogLevel get logLevel => settings.logLevel;