+++ packages/talker_riverpod_logger/lib/riverpod_logs.dart
@@ -27,7 +27,7 @@ class RiverpodAddLog extends TalkerLog {
   final TalkerRiverpodLoggerSettings settings;
 
   @override
-  String get key => TalkerLogType.riverpodAdd.key;
+  String get key => TalkerKey.riverpodAdd;
 
   @override
   String generateTextMessage({
@@ -62,7 +62,7 @@ class RiverpodUpdateLog extends TalkerLog {
   final TalkerRiverpodLoggerSettings settings;
 
   @override
-  String get key => TalkerLogType.riverpodUpdate.key;
+  String get key => TalkerKey.riverpodUpdate;
 
   @override
   String generateTextMessage({
@@ -95,7 +95,7 @@ class RiverpodDisposeLog extends TalkerLog {
   final TalkerRiverpodLoggerSettings settings;
 
   @override
-  String get key => TalkerLogType.riverpodDispose.key;
+  String get key => TalkerKey.riverpodDispose;
 
   @override
   String generateTextMessage({
@@ -128,7 +128,7 @@ class RiverpodFailLog extends TalkerLog {
   final TalkerRiverpodLoggerSettings settings;
 
   @override
-  String? get key => TalkerLogType.riverpodFail.key;
+  String? get key => TalkerKey.riverpodFail;
 
   @override
   String generateTextMessage({